django_api/app/pyproject.toml
[tool.poetry]
name = "docker-compose"
version = "0.1.0"
description = ""
authors = ["dosxana <dosxana@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-debug-toolbar = "^4.2.0"
django-split-settings = "^1.2.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
flake8 = "^7.0.0"
sqlparse = "^0.4.4"
install = "^1.3.5"
gunicorn = "^21.2.0"
daphne = "^4.0.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
drf-yasg = "^1.21.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
